Add TleRepositry().find()
